+++ .eslintrc.js
@@ -2,7 +2,7 @@ module.exports = {
   env: {
     browser: true,
     commonjs: true,
-    es6: true,
+    es2022: true,
     'jest/globals': true,
     mongo: true,
   },