+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 [BREAKING] change i18n middleware to internal one. Nothing should be affected
 [BREAKING] now validation of request splitted between request and query
 [BREAKING] supportedLngs option added to i18n config
+[BREAKING] email inliner now looking for src/services/messaging/email/resources folder instead of 'build' folder.
 
 [BREAKING] Mongoose v7. https://mongoosejs.com/docs/migrating_to_7.html